requirements.txt
streamlit
firebase-admin
python-dotenv
google-generativeai
requests
pandas
